+++ scripts/start.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
 
 defaultDir=/opt/open-webui
+condaSh=/$defaultDir/miniconda3/etc/profile.d/conda.sh
+condaEnv=$defaultDir/config/conda/open-webui
 openWebuiStart=$defaultDir/open-webui/backend/start.sh
-condaDir=$defaultDir/config/conda
-condaEnv=$condaDir/open-webui
 
-#source /etc/profile.d/conda.sh
-source /usr/etc/profile.d/conda.sh
+source $condaSh
 conda activate $condaEnv
 $openWebuiStart